requirements.txt
streamlit
langchain_openai
python-dotenv
crewai
pypdf